pyproject.toml
[project]
name = "fix-echo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.33.0",
    "ipython>=9.3.0",
    "openai>=1.84.0",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
